refactor: run formatter and simplify ban gateway
