+++ src/bans/interfaces/bans-gateway.interface.ts
@@ -1,0 +1,7 @@
+export enum BanGatewayRoom {
+  NEW_BAN_NOTIFICATION = 'new-ban-notification',
+}
+
+export enum BanGatewayEvent {
+  NEW_BANS = 'bans',
+}
