+++ src/bans/interfaces/bans.interface.ts
