+++ src/bans/services/bans.service.ts
@@ -2,7 +2,7 @@ import { forwardRef, Inject, Injectable, Logger } from '@nestjs/common';
 import { Cron, SchedulerRegistry } from '@nestjs/schedule';
 import { RglService } from '../../rgl/rgl.service';
 import { CronNames } from '../../enums/crons.enum';
-import { Ban } from '../bans.interface';
+import { Ban } from '../interfaces/bans.interface';
 import { CacheService } from 'src/cache/cache.service';
 import { Events } from 'src/events/events';
 import { StartupService } from 'src/startup/services/startup.service';
