+++ src/startup/services/startup.service.ts
@@ -41,7 +41,9 @@ export class StartupService implements OnModuleInit {
       await this.startupSchema.findOneAndUpdate({}, { startingBan: steamId });
     } catch (err) {
       this.STARTING_BAN = steamId;
-      return this.logger.error(`Could not save steamId (${steamId}) to MongoDB.`);
+      return this.logger.error(
+        `Could not save steamId (${steamId}) to MongoDB.`,
+      );
     }
     this.STARTING_BAN = steamId;
     this.logger.debug('Successfully set new starting ban.');
